requirements.txt
websocket-client
requests
numpy
matplotlib
# sudo apt install python3-tk
